==> submissions/RU_Y2_HiFi/1-wrangle-primrose.ipynb <==
import pandas as pd
import glob
import numpy as np

# Function to add single quotation marks around each element and convert list to string
def add_quotes_and_convert_to_string(lst):
    return str([f"{item}" for item in lst])

ru_y2_df = pd.read_csv('RU_Y2_HIFI.csv',header=None)
ru_y2_df.rename(columns={0:'hifi_reads'}, inplace=True)
ru_y2_df = ru_y2_df[ru_y2_df['hifi_reads'].str.contains('hifi_reads')]
# ru_y2_df['sample_id'] = [sample.split('/')[-2] + '_' + sample.split('/')[-1] for sample in ru_y2_df['hifi_reads'].tolist()]


sample_id_list = []
for data in ru_y2_df['hifi_reads'].tolist():
	sample_id = data.split('/')[-2]
	if sample_id == 'PacBio_HiFi':
		sample_id_list.append(data.split('/')[-3])
	else:
		sample_id_list.append(sample_id)
ru_y2_df['sample_id'] = sample_id_list

ru_y2_df= ru_y2_df[~ru_y2_df['sample_id'].isin(['HG04187'])] # removing sample because the files are missing 5mc
# ru_y2_df['sample_id'] = ru_y2_df['sample_id'] + '_' +  [sample.split('/')[-1] for sample in ru_y2_df['hifi_reads'].tolist()]


primrose_df = pd.DataFrame(ru_y2_df['hifi_reads'].tolist())
primrose_df.columns = ['input_bams']
primrose_df['sample_id'] = ru_y2_df['sample_id'] 

# primrose_df[['sample_id','input_bams']].to_csv('primrose/primrose_sample_data.csv',index=False)

primrose_df

primrose_df = primrose_df.groupby('sample_id')['input_bams'].apply(list).reset_index()
primrose_df['input_bams'] = np.array([add_quotes_and_convert_to_string(data) for data in primrose_df['input_bams'].values])
primrose_df[['sample_id', 'input_bams']].to_csv('primrose_sample_data.csv',index=False)

primrose_df

sample_pull = []
other_reads = []
for sample in [sample.split('_')[0] for sample in ru_y2_df['sample_id'].tolist()]:
	other_reads_path = glob.glob('/private/groups/hprc/human-pangenomics/1000G_local/child/cram_to_bam/'+sample+'/*/*/*/*')
	if other_reads_path == []:

		other_reads_path = glob.glob('/private/groups/hprc/1000G_local/child/fastq/' + sample + '.*')
		if other_reads_path == []:
			sample_pull.append(sample)
		else:
			other_reads.append([other_reads_path[0]])
	else:
		other_reads.append([other_reads_path[0]])
ru_y2_df['other_reads'] = other_reads
ru_y2_df['perform_methylation_check'] = 'true'


ru_y2_df[['sample_id','hifi_reads', 'other_reads', 'perform_methylation_check']].to_csv('hifi_qc/hifi_qc_sample_files.csv', index=False)
