==> submissions/RU_Y2_HiFi/2-evaluate-primrose.ipynb <==
import pandas as pd
import glob
import os
import shutil

bam_files = glob.glob('primrose/HG*/analysis/primrose_outputs/*/*')



target_base_dir = 'upload-primrose/'

def restructure_files(bam_files, target_base_dir):
    for bam_file in bam_files:
         # Extract parts from the file path
        parts = bam_file.split('/')
        if '.bam.5mc.hifi_reads.bam' in bam_file:
            bam_filename = parts[-1]
            bam_filename = bam_filename.replace('.bam.5mc.hifi_reads.bam', '.5mc.hifi_reads.bam')
        else:
            bam_filename = parts[-1]

       
        sample_id = parts[-5].split('_')[0]
        run_id = parts[-4]
        
        new_dir = os.path.join(target_base_dir, sample_id, 'primrose')
        os.makedirs(new_dir, exist_ok=True)

        new_file_path = os.path.join(new_dir, bam_filename)
        shutil.move(bam_file, new_file_path)
        print(f"Moved {bam_file} to {new_file_path}")


# Run the restructuring function
restructure_files(bam_files, target_base_dir)

ssds staging upload \
--deployment default \
--submission-id 746FF75B-3C32-42AF-92BB-290BA92CF89A  \
upload-primrose
