==> submissions/RU_Y2_HiFi/4-evaluate-hifi_qc.ipynb <==
import pandas as pd

hifi_qc = pd.read_csv('/private/groups/hprc/human-pangenomics/HPRC_metadata/submissions/RU_Y2_HiFi/hifi_qc/hifi_qc_samples_output_table.csv')
hifi_meta = pd.read_csv('HPRC_RU_Y2_Sequel_Metadata_PacBio_HiFi_Submission.csv')

hifi_dict = {data[0]:pd.read_csv(data[1], sep='\t') for data in hifi_qc[['sample_id', 'hifi_qc_summary']].values}

hifi_df = pd.concat(hifi_dict.values())
hifi_df['sample_ID'] = [sample_id.split('_')[0] for sample_id in hifi_df['filename'].tolist()]
# hifi_df.drop(columns=['bam_file'],inplace=True)
# hifi_df.rename(columns={'bam_file':'filename'},inplace=True)

hifi_df = hifi_df[['sample_ID',
'filename',
'bam_file',
 'total_reads',
 'total_bp',
 'total_Gbp',
 'min',
 'max',
 'mean',
 'quartile_25',
 'quartile_50',
 'quartile_75',
 'N25',
 'N50',
 'N75',
 'ntsm_score',
 'ntsm_result',
 'MM_tag',
 'ML_tag']]

hifi_meta['flow_cell'] = hifi_meta['filename']
hifi_df['flow_cell'] = [flow_cell.split('.5mc.')[0].split('_',1)[1] for flow_cell in hifi_df['filename'].tolist()]

hifi_df.to_csv('RU_Y2_HIFI_post_sra_metadata.csv',index=False)
