==> submissions/RU_Y2_HiFi/6-metadata.ipynb <==
import collections
import pandas as pd
import mechanicalsoup

ru_y2_metadata = pd.read_csv('HPRC_RU_Y2_Sequel_Metadata_PacBio_HiFi_Submission_Initial.csv')
ru_y2_metadata['flow_cell'] = [flow_cell.split('.')[0] for flow_cell in ru_y2_metadata['filename'].tolist()]
ru_y2_metadata.shape

ru_y2_qc = pd.read_csv('RU_Y2_HIFI_post_sra_metadata.csv')
ru_y2_qc.rename(columns={'filename':'sample_ID_filename', 'bam_file':'filename'},inplace=True)
print(ru_y2_qc.shape)


# 18 files are not present in qc sample: HG04187
ru_y2_metadata[~ru_y2_metadata['flow_cell'].isin(ru_y2_qc['flow_cell'].tolist())].shape 

# all flow cells are present between qc and metadata
ru_y2_qc = ru_y2_qc[ru_y2_metadata['flow_cell'].isin(ru_y2_metadata['flow_cell'].tolist())]

ru_y2_df = pd.merge( ru_y2_metadata, ru_y2_qc[['flow_cell','filename']], on='flow_cell', how='inner')
ru_y2_df.drop('filename_x', axis=1, inplace=True)
ru_y2_df.rename(columns={'filename_y':'filename'}, inplace=True)

ru_y2_setup = pd.read_csv('RU_Y2_HIFI.csv', header=None)

ru_y2_setup = ru_y2_setup[ru_y2_setup[0].str.contains('primrose')]
ru_y2_setup['filename'] = [filename.split('/')[-1] for filename in ru_y2_setup[0].tolist()]
ru_y2_setup.rename(columns={0:'s3-submission'},inplace=True)

# update primrose working paths
transfer_list = []
for data in ru_y2_setup['s3-submission'].tolist():
	path = data.split('/746FF75B-3C32-42AF-92BB-290BA92CF89A--RU_Y2_HIFI')[1].split('primrose')
	path.insert(1, 'raw_data/PacBio_HiFi/primrose')
	path = ''.join(path)
	release = 's3://human-pangenomics/working/HPRC' + path
	transfer_list.append(release)
ru_y2_setup['s3-working'] = transfer_list
ru_y2_setup[['s3-submission','s3-working']].to_csv('RU_Y2_HiFi.transfer_ec2.csv',index=False, header=None)


ru_y2_df = pd.merge(ru_y2_df,ru_y2_setup, on='filename', how='inner')

ru_y2_df.to_csv('HPRC_RU_Y2_Sequel_Metadata_PacBio_HiFi_Submission_Final.csv',index=False)

ru_y2_df = pd.read_csv('HPRC_RU_Y2_Sequel_Metadata_PacBio_HiFi_Submission_Final.csv')

ru_y2_df.head()

# Create BioSample Ids

# biosample check
hprc_biosample = pd.read_csv('/private/groups/hprc/human-pangenomics/documentation-metadata/HPRC_metadata/data/production/hprc-production-biosample-table.tsv', sep='\t')
hprc_biosample.rename(columns={'Sample':'sample_ID'},inplace=True)

ncbi_biosample_df = pd.read_csv('ncbi-biosample-07.31.24.tsv',sep='\t')
ncbi_biosample_df.rename(columns={'BioSample.name':'sample_ID'},inplace=True)

assert set(ru_y2_df[~ru_y2_df['sample_ID'].isin(ncbi_biosample_df.sample_ID.tolist())].sample_ID.tolist()) == set(ru_y2_df[~ru_y2_df['sample_ID'].isin(hprc_biosample.sample_ID.tolist())].sample_ID.tolist())

# Create a MechanicalSoup browser object
manual_resolve = []
browser = mechanicalsoup.Browser()
sample_collection_dict = {sample : {"tissue": None,
                                    "geo_loc_name": None,
                                    "sex": None
                                   } for sample in ru_y2_df[~ru_y2_df['sample_ID'].isin(hprc_production.sample_ID.tolist())].sample_ID.tolist()}

sample_collection_dict

for sample in sample_collection_dict.keys():
    
    url = 'https://catalog.coriell.org/0/Sections/Search/Sample_Detail.aspx?Ref='+sample+'&Product=DNA' # +'&product=CC'
    page = browser.get(url)
    
    form = page.soup.select('form')[0]

    try:
        country_span = form.select_one('#lblCountry').text
        cell_type_span = form.select_one('#lblCell_Type').text
        sex_span = form.select_one('#lblGender').text
        sample_collection_dict[sample]['tissue'] = cell_type_span
        sample_collection_dict[sample]['geo_loc_name'] = country_span
        sample_collection_dict[sample]['sex'] = sex_span
    except AttributeError:
        print(sample)

# create biosample identifiers and import
sample_collection_dict['HG01887']['Accession'] = 'SAMN42953248'
sample_collection_dict['HG02656']['Accession'] = 'SAMN42953249'
sample_collection_dict['HG03825']['Accession'] = 'SAMN42953250'


sample_collection_df = pd.DataFrame.from_dict(sample_collection_dict,orient='index')

sample_collection_df

hprc_production = pd.read_csv('hprc-production-07.31.24.tsv',sep='\t',header=1)

# hprc_production[hprc_production['ChildID'].isin(['HG01626'])]

hprc_production[hprc_production['ChildID'].isin(sample_collection_df.index.tolist())]

ru_y2_bio_obj = pd.read_csv('RU_Y2_HIFI_BioSampleObject.tsv',sep='\t')
ru_y2_bio_obj.rename(columns={'Sample Name': 'ChildID'},inplace=True)

pd.merge(ru_y2_bio_obj,hprc_production[['ChildID', 'N', 'Subpopulation', 'Superpopulation', 'Production Year']], on='ChildID')

# update biosample tracking
hprc_biosample

# AWS Release Check

# Check sampleID-flow cell id combo in primrose with top level 
aws_df = pd.read_csv('RU_Y2_HIFI.csv', header=None)

aws_df['flow_cell'] = [flow_cell.split('/')[-1].split('.')[0] for flow_cell in aws_df[0].tolist()]
aws_df['sample_id'] = [flow_cell.split('/')[-2] for flow_cell in aws_df[0].tolist()]

aws_df = aws_df[~aws_df['sample_id'].isin(['746FF75B-3C32-42AF-92BB-290BA92CF89A--RU_Y2_HIFI'])]
aws_df = aws_df[~aws_df[0].str.contains('HG04187')]

aws_primrose = aws_df[aws_df['sample_id'].isin(['primrose'])]
aws_primrose['sample_id'] = [sample.split('/')[-3] for sample in aws_primrose[0].tolist()]

aws_df = aws_df[~aws_df['sample_id'].isin(['primrose'])]

len([sample for sample in set(aws_primrose['flow_cell'].tolist()) if sample not in set(aws_df['flow_cell'].tolist())])

len([sample for sample in set(aws_df['flow_cell'].tolist()) if sample not in set(aws_primrose['flow_cell'].tolist())])

print(len(set(aws_df['flow_cell'].tolist())))
print(len(set(aws_df['flow_cell'].tolist())))

set(aws_df['flow_cell'].tolist()) == set(aws_df['flow_cell'].tolist())

aws_primrose['sampleID-flowcellID'] = aws_primrose['sample_id'] + '_' + aws_primrose['flow_cell']

aws_df.head()

aws_df = aws_df[~aws_df['sample_id'].isin(['PacBio_HiFi'])]

aws_df['sampleID-flowcellID'] = aws_df['sample_id'] + '_' + aws_df['flow_cell']

len(collections.Counter(aws_primrose['sampleID-flowcellID'].tolist()).keys())

len(collections.Counter(aws_df['sampleID-flowcellID'].tolist()))

for data in list(collections.Counter(aws_df['sampleID-flowcellID'].tolist()).keys()):
	if data not in list(collections.Counter(aws_primrose['sampleID-flowcellID'].tolist()).keys()):
		print(data)

for data in list(collections.Counter(aws_primrose['sampleID-flowcellID'].tolist()).keys()):
	if data not in list(collections.Counter(aws_df['sampleID-flowcellID'].tolist()).keys()):
		print(data)

ssds staging release     --deployment default     --submission-id 746FF75B-3C32-42AF-92BB-290BA92CF89A    --transfer-csv RU_Y2_HiFi.transfer_ec2.csv     &> RU_Y2_HiFi.transfer_aws.stdout


# SRA

ru_y2_metadata = pd.read_csv('/private/groups/hprc/human-pangenomics/HPRC_metadata/submissions/RU_Y2_HiFi/HPRC_RU_Y2_Sequel_Metadata_PacBio_HiFi_Submission_Final.csv')

ru_y2_metadata.head()

ru_y2_metadata = ru_y2_metadata[['sample_ID', 'filename','library_ID', 'library_strategy','library_source','library_selection', 'library_layout',
				'platform','instrument_model','shear_method', 'size_selection', 'ccs_algorithm', 'polymerase_version',
				'seq_plate_chemistry_version', 'generator_facility','generator_contact','filetype']]
ru_y2_metadata['assembly'] = 'unaligned'

ru_y2_sra = pd.merge(ru_y2_metadata, hprc_biosample[['sample_ID','Accession']], on='sample_ID', how='inner')

ru_y2_sra['title'] = ['PacBio HiFi sequencing of ' + sample_id for sample_id in ru_y2_sra['sample_ID'].tolist()]
ru_y2_sra['design_description'] = 'HiFi sequencing of 20kb fractionated gDNA'

ru_y2_sra.rename(columns={'Accession':'biosample_accession'},inplace=True)

ru_y2_sra['flow_cell'] = [flow_cell.split('.')[0] for flow_cell in ru_y2_sra['filename'].tolist()]
ru_y2_sra['library_ID'] = ru_y2_sra['library_ID'] + '_' + ru_y2_sra['flow_cell']

ru_y2_sra.drop(columns=['sample_ID','flow_cell'],inplace=True)

ru_y2_sra['instrument_model'] = 'Sequel II'

ru_y2_sra.to_csv('sra_biosample_RU_Y2_metadata.tsv',sep='\t', index=False)

# Add RU Y2 SRA Identifiers
After submitting to NCBI

import pandas as pd
ru_y2_sra = pd.read_csv('sra_biosample_RU_Y2_metadata.tsv', sep='\t')
ru_y2_sra_ids = pd.read_csv('/private/groups/hprc/human-pangenomics/documentation-metadata/HPRC_metadata/submissions/RU_Y2_HiFi/metadata-14647736-processed-ok.tsv',sep='\t')


# library_ID joining
ru_y2_sra = pd.merge(ru_y2_sra_ids[['accession','study','bioproject_accession','library_ID']], ru_y2_sra, on='library_ID',how='inner')

final_ids = list(pd.read_csv('/private/groups/hprc/human-pangenomics/documentation-metadata/HPRC_metadata/submissions/RU_Y2_HiFi/HPRC_PacBio_HiFi.csv'))

ru_y2_sra
