==> submissions/RU_Y2_HiFi/7-add-datatable-HPRC_PacBio_HiFi.ipynb <==
import numpy as np
import pandas as pd
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

sra_identifiers = pd.read_csv('/private/groups/hprc/human-pangenomics/documentation-metadata/HPRC_metadata/submissions/RU_Y2_HiFi/metadata-14647736-processed-ok.tsv',sep='\t')

post_sra = pd.read_csv('/private/groups/hprc/human-pangenomics/documentation-metadata/HPRC_metadata/submissions/RU_Y2_HiFi/RU_Y2_HIFI_post_sra_metadata.csv')

sra_identifiers['bam_file'] = sra_identifiers['filename']

post_sra[~post_sra['bam_file'].isin(sra_identifiers.bam_file.tolist())]

sra_identifiers[sra_identifiers['bam_file'].isin(post_sra.bam_file.tolist())].shape

ru_y2_hifi = pd.merge(sra_identifiers, post_sra, on='bam_file',how='inner')

ru_y2_hifi.to_csv('RU_Y2_HiFi_final_table.csv',index=False)

hprc_pacbio_hifi = pd.read_csv('/private/groups/hprc/human-pangenomics/documentation-metadata/HPRC_metadata/data/hprc-data-explorer-tables/HPRC_PacBio_HiFi.tsv',sep='\t')

hprc_pacbio_hifi['MM_tag'] = False

# ru_y2_hifi[['filename_x', 'filename_y']]
ru_y2_hifi.drop('filename_y', axis=1,inplace=True)
ru_y2_hifi.rename(columns={'filename_x':'filename'},inplace=True)

ru_y2_hifi_release = pd.read_csv('/private/groups/hprc/human-pangenomics/documentation-metadata/HPRC_metadata/submissions/RU_Y2_HiFi/RU_Y2_HiFi.transfer_ec2.csv',header=None)


ru_y2_hifi_release['filename'] = [filename.split('/')[-1] for filename in ru_y2_hifi_release[1].tolist()]
ru_y2_hifi_release.rename(columns={1:'path'},inplace=True)

ru_y2_hifi = pd.merge(ru_y2_hifi, ru_y2_hifi_release[['filename','path']], on='filename', how='inner')

ru_y2_hifi['data_type'] = 'unaligned reads'

ru_y2_hifi['MM_tag'] = True
ru_y2_hifi['DeepConsensus_version'] = np.nan
ru_y2_hifi['notes'] = np.nan

hprc_pacbio_hifi = pd.concat([hprc_pacbio_hifi, ru_y2_hifi[list(hprc_pacbio_hifi)]])

# List of example columns (from the metadata sheet)

# Check for missing columns in the new data
missing_columns = [col for col in list(hprc_pacbio_hifi) if col not in ru_y2_hifi.columns]

# Check for extra columns in the new data
extra_columns = [col for col in ru_y2_hifi.columns if col not in list(hprc_pacbio_hifi)]

missing_columns

hprc_pacbio_hifi.to_csv('/private/groups/hprc/human-pangenomics/documentation-metadata/HPRC_metadata/data/hprc-data-explorer-tables/HPRC_PacBio_HiFi.tsv',sep='\t', index=False)
